Renamed classes to Superhero
