+++ app/src/main/java/com/example/gitrobotdreamsleesonsnewhomeasks/ApiInterface.kt
@@ -5,5 +5,5 @@ import retrofit2.http.GET
 
 interface ApiInterface {
    @GET("/get_memes")
-    fun getMemes():Single<MemesResponse>
+    fun getSuperHeroes():Single<SuperHeroDataResponse>
 }
