+++ app/src/main/java/com/example/gitrobotdreamsleesonsnewhomeasks/MainActivity.kt
@@ -2,12 +2,9 @@ package com.example.gitrobotdreamsleesonsnewhomeasks
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.ListView
 import android.widget.Toast
-import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import io.reactivex.Scheduler
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
 
@@ -17,13 +14,13 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.activity_main)
         val recyclerView:RecyclerView = findViewById(R.id.recyclerView)
         val api = ApiClient.client.create(ApiInterface::class.java)
-        api.getMemes()
+        api.getSuperHeroes()
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe ({
                   if(it.data.memes != null){
                       val items = it.data.memes
-                      val myAdapter = RecyclerViewAdapter(items as MutableList<Meme>){
+                      val myAdapter = SuperheroViewAdapter(items as MutableList<SuperHero>){
                           Toast.makeText(this, "${it} clicked", Toast.LENGTH_SHORT).show()
                       }
                       recyclerView.adapter =myAdapter
@@ -36,6 +33,6 @@ class MainActivity : AppCompatActivity() {
     }
 }
 
-data class MemesResponse(val data:Data)
-data class Data(val memes:List<Meme>)
-class Meme(val name:String, val url:String)
+data class SuperHeroDataResponse(val data:Data)
+data class Data(val memes:List<SuperHero>)
+class SuperHero(val name:String, val url:String)
